Download screenshot file in background script and add save as option (#60)
* Download screenshot file in background script

Instead of adding an auto-clicked hyperlink containing screenshot data to
dowload it, pass the blob to background script and use browser.downloads
interface to do so.  It avoids memory leak as the browser does not really free
up the memory of the hyperlink even if removed from DOM.

* Add save as option when downloading screenshot

Using browser.download() allows such a feature. Add a preference option to open
the file chooser dialog if enabled.